Use router query for data fetching
