+++ src/pages/Home/Home.jsx
@@ -1,32 +1,22 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+import { useLocation } from 'react-router-dom';
 
-import { Container, Spinner } from '../../components/UI';
+import { Container } from '../../components/UI';
 import VideoList from '../../components/VideoList';
 
 import { useYoutubeData } from '../../hooks';
 
 function HomePage() {
-  const { loading, videos, error } = useYoutubeData();
-  console.log(error);
+  const { videos, search: doSearch } = useYoutubeData();
+  const { pathname, search } = useLocation();
+  const params = new URLSearchParams(search);
+  const q = params.get('q');
 
-  if (loading) {
-    return (
-      <div
-        style={{
-          display: 'flex',
-          justifyContent: 'center',
-          alignItems: 'center',
-          flex: 1,
-        }}
-      >
-        <Spinner />
-      </div>
-    );
-  }
-
-  if (error) {
-    return <span>{JSON.stringify(error, null, 2)}</span>;
-  }
+  useEffect(() => {
+    if (pathname === '/') {
+      doSearch({ query: q });
+    }
+  }, [doSearch, pathname, q]);
 
   if (!videos) {
     return null;
